{{cookiecutter.github_repository_name}}/requirements/base.txt
# Django
Django==1.9.8

# Configuration
django-environ==0.4.0

# For the persistence stores
psycopg2==2.6.2
dj-database-url==0.4.1

# Models
django-model-utils==2.5.2

# Rest apis
djangorestframework==3.4.6
Markdown==2.6.6
django-filter==0.14.0
django-crispy-forms

# Static files
whitenoise==3.2.1

# Media files
boto==2.42.0
django_unique_upload==0.2.1

# Image manipulation
django-versatileimagefield==1.6

# For asynchronous queuing
django-rq==0.9.2

# Time zones support
pytz==2016.6.1

# Ali oss SDK
oss2
# django-aliyun-oss2-storage
git+https://github.com/daimon99/django-aliyun-oss2-storage

# Swagger(api文档模块)
django-rest-swagger

django-import-export

# 微信与加解密
wechatpy[cryptography]